+++ src/main/java/oci_utils/model/BaseOciDataList.java
@@ -1,5 +1,7 @@
 package oci_utils.model;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
@@ -9,4 +11,8 @@ import java.util.List;
 @Accessors(chain = true)
 public class BaseOciDataList<T extends BaseOciEntity> {
     List<T> data;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonProperty("opc-next-page")
+    String nextPage;
 }