+++ src/main/java/oci_utils/model/SessionItem.java
@@ -4,23 +4,22 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
 import java.time.Instant;
 
+@EqualsAndHashCode(callSuper = true)
 @Data
 @Accessors(chain = true)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class SessionItem {
+@Dto
+public class SessionItem extends BaseOciEntity {
     String bastionId;
     String bastionName;
     @JsonProperty("bastion-public-host-key-info")
     JsonNode bastionPublicHostKeyInfo;
     @JsonProperty("bastion-user-name")
     String bastionUserName;
-    @JsonProperty("display-name")
-    String displayName;
-    String id;
     @JsonProperty("key-details")
     KeyDetails keyDetails;
     @JsonProperty("key-type")